Update heartbeat rate/measurment rate, heartbeat rate type is uint64, default value is 600s. measurement rate default value is 120s

+++ components/heartbeat/heartbeat.h
@@ -33,7 +33,7 @@ void heartbeat_init();
 **Description**
     Set the internal value for the heartbeat rate.
 */
-void heartbeat_set_period(uint8_t period);
+void heartbeat_set_period(uint64_t period);
 
 /** Start the heartbeat timer.
